Move username validation logic and its tests
