+++ src/client/UsernameInput.ts
@@ -3,10 +3,8 @@ import { customElement, property, state } from "lit/decorators.js";
 import { v4 as uuidv4 } from "uuid";
 import { translateText } from "../client/Utils";
 import { UserSettings } from "../core/game/UserSettings";
-import {
-  MAX_USERNAME_LENGTH,
-  validateUsername,
-} from "../core/validations/username";
+import { MAX_USERNAME_LENGTH } from "../core/validations/username";
+import { validateUsername } from "./validation/username";
 
 const usernameKey: string = "username";
