+++ src/client/validation/username.ts
@@ -1,0 +1,45 @@
+import {
+  MAX_USERNAME_LENGTH,
+  MIN_USERNAME_LENGTH,
+  validPattern,
+} from "../../core/validations/username";
+import { translateText } from "../Utils";
+
+export function validateUsername(username: string): {
+  isValid: boolean;
+  error?: string;
+} {
+  if (typeof username !== "string") {
+    return { isValid: false, error: translateText("username.not_string") };
+  }
+
+  if (username.length < MIN_USERNAME_LENGTH) {
+    return {
+      isValid: false,
+      error: translateText("username.too_short", {
+        min: MIN_USERNAME_LENGTH,
+      }),
+    };
+  }
+
+  if (username.length > MAX_USERNAME_LENGTH) {
+    return {
+      isValid: false,
+      error: translateText("username.too_long", {
+        max: MAX_USERNAME_LENGTH,
+      }),
+    };
+  }
+
+  if (!validPattern.test(username)) {
+    return {
+      isValid: false,
+      error: translateText("username.invalid_chars", {
+        max: MAX_USERNAME_LENGTH,
+      }),
+    };
+  }
+
+  // All checks passed
+  return { isValid: true };
+}
