+++ src/core/validations/username.ts
@@ -7,7 +7,6 @@ import {
   resolveLeetSpeakTransformer,
   skipNonAlphabeticTransformer,
 } from "obscenity";
-import { translateText } from "../../client/Utils";
 import { simpleHash } from "../Util";
 
 const matcher = new RegExpMatcher({
@@ -22,7 +21,7 @@ const matcher = new RegExpMatcher({
 export const MIN_USERNAME_LENGTH = 3;
 export const MAX_USERNAME_LENGTH = 27;
 
-const validPattern = /^[a-zA-Z0-9_[\] 🐈🍀üÜ]+$/u;
+export const validPattern = /^[a-zA-Z0-9_[\] 🐈🍀üÜ]+$/u;
 
 const shadowNames = [
   "NicePeopleOnly",
@@ -43,45 +42,6 @@ export function fixProfaneUsername(username: string): string {
 
 export function isProfaneUsername(username: string): boolean {
   return matcher.hasMatch(username);
-}
-
-export function validateUsername(username: string): {
-  isValid: boolean;
-  error?: string;
-} {
-  if (typeof username !== "string") {
-    return { isValid: false, error: translateText("username.not_string") };
-  }
-
-  if (username.length < MIN_USERNAME_LENGTH) {
-    return {
-      isValid: false,
-      error: translateText("username.too_short", {
-        min: MIN_USERNAME_LENGTH,
-      }),
-    };
-  }
-
-  if (username.length > MAX_USERNAME_LENGTH) {
-    return {
-      isValid: false,
-      error: translateText("username.too_long", {
-        max: MAX_USERNAME_LENGTH,
-      }),
-    };
-  }
-
-  if (!validPattern.test(username)) {
-    return {
-      isValid: false,
-      error: translateText("username.invalid_chars", {
-        max: MAX_USERNAME_LENGTH,
-      }),
-    };
-  }
-
-  // All checks passed
-  return { isValid: true };
 }
 
 export function sanitizeUsername(str: string): string {
